备份/requirements.txt
# ==================== Python依赖包 ====================
# Telegram Bot API
pyrogram>=2.0.0
tgcrypto>=1.2.0

# Firebase
firebase-admin>=6.0.0

# 异步支持
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0

# 数据处理
python-dotenv>=1.0.0

# 日志和工具
colorlog>=6.7.0

# 进程管理
psutil>=5.9.0
google-generativeai>=0.7.0
